(chore) Configure UpdateSchools job to run via Sidekiq
