+++ lib/tasks/data.rake
@@ -3,7 +3,7 @@ namespace :data do
   namespace :schools do
     task import: :environment do
       Rails.logger.debug("Running school import task in #{Rails.env}")
-      UpdateSchoolsDataFromSourceJob.new.perform
+      UpdateSchoolsDataFromSourceJob.perform_later
     end
   end
 end
